pridany zeme s vlajkami, do databaze naimprotovane ligy a zeme z basketbalu

+++ src/main/java/cz/vse/aplikaceprosledovanivysledkusportovnichzapasu/entity/Country.java
@@ -10,7 +10,7 @@ public class Country {
     @Column(name = "id", nullable = false)
     private Long id;
     private String name;
-    private Long externalId;
+    private int externalId;
     private String flag;
     private String sport;
 }
